requirements.txt
streamlit
streamlit-extras
tiktoken
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
docx==0.2.4
executing==0.8.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
ipykernel==6.25.0
ipython==8.12.0
jedi==0.18.1
jupyter_client==8.1.0
jupyter_core==5.3.0
langchain==0.0.257
langsmith==0.0.19
lxml==4.9.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.0
pure-eval==0.2.2
pydantic==1.10.12
PyDocX==0.9.10
Pygments==2.15.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
setuptools==68.0.0
six==1.16.0
spyder-kernels==2.4.3
SQLAlchemy==2.0.19
stack-data==0.2.0
tenacity==8.2.2
tornado==6.3.2
tqdm==4.65.1
traitlets==5.7.1
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.5
wheel==0.38.4
XlsxWriter==3.1.2
yarl==1.9.2

